refactor: 패키지 구조 변경
- 모델이라는 이름 대신 도메인을 사용합니다.
- 표현, 도메인, 영속화 계층으로 분리합니다.
