+++ src/main/java/roomescape/domain/exception/DomainExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.domain.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
