+++ src/main/java/roomescape/domain/exception/ReservationException.java
@@ -1,4 +1,4 @@
-package roomescape.exception;
+package roomescape.domain.exception;
 
 public class ReservationException extends IllegalArgumentException {
