+++ src/main/java/roomescape/persist/entity/ReservationTimeEntity.java
@@ -1,4 +1,4 @@
-package roomescape.entity;
+package roomescape.persist.entity;
 
 import java.util.Objects;
