+++ src/main/java/roomescape/persist/repository/H2ReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.repository;
+package roomescape.persist.repository;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -9,11 +9,11 @@ import java.util.Map;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.entity.ReservationEntity;
-import roomescape.entity.ReservationTimeEntity;
-import roomescape.model.Reservation;
-import roomescape.model.ReservationDate;
-import roomescape.model.ReservationTime;
+import roomescape.persist.entity.ReservationEntity;
+import roomescape.persist.entity.ReservationTimeEntity;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationDate;
+import roomescape.domain.ReservationTime;
 
 @Repository
 public class H2ReservationRepository implements ReservationRepository {
