+++ src/main/java/roomescape/persist/repository/ReservationRepository.java
@@ -1,7 +1,7 @@
-package roomescape.repository;
+package roomescape.persist.repository;
 
 import java.util.List;
-import roomescape.model.Reservation;
+import roomescape.domain.Reservation;
 
 public interface ReservationRepository {
