+++ src/main/java/roomescape/presentation/controller/ReservationApiController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.presentation.controller;
 
 import jakarta.validation.Valid;
 import java.util.List;
@@ -9,14 +9,14 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationTimeResponseDto;
-import roomescape.model.ReservationDate;
-import roomescape.model.ReservationTime;
-import roomescape.repository.ReservationRepository;
-import roomescape.model.Reservation;
-import roomescape.dto.ReservationRequestDto;
-import roomescape.dto.ReservationResponseDto;
-import roomescape.repository.ReservationTimeRepository;
+import roomescape.presentation.dto.ReservationTimeResponseDto;
+import roomescape.domain.ReservationDate;
+import roomescape.domain.ReservationTime;
+import roomescape.persist.repository.ReservationRepository;
+import roomescape.domain.Reservation;
+import roomescape.presentation.dto.ReservationRequestDto;
+import roomescape.presentation.dto.ReservationResponseDto;
+import roomescape.persist.repository.ReservationTimeRepository;
 
 @RestController
 @RequestMapping("/reservations")
