+++ src/test/java/roomescape/domain/ReservationTest.java
@@ -1,4 +1,4 @@
-package roomescape.model;
+package roomescape.domain;
 
 import static org.assertj.core.api.Assertions.assertThatCode;
 
@@ -8,7 +8,7 @@ import java.time.LocalTime;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
-import roomescape.exception.ReservationException;
+import roomescape.domain.exception.ReservationException;
 
 class ReservationTest {
